+++ frontend/messages/en.json
@@ -1,0 +1,98 @@
+{
+    "HomePage": {
+      "title": "MAESTRE",
+      "subtitle": "Reimagining education with intelligence",
+      "main_button": "Discover the AI tools",
+      "main_description": "Maestre combines cutting-edge artificial intelligence with pedagogical expertise to create tools that enhance teaching and elevate student learning experiences",
+      "loading": "Loading content...",
+
+      "feature_title_1": "Step 1: Create & Manage Classes",
+      "feature_description_1": "Set up your virtual classroom environment by creating and organizing classes for different subjects and student groups.",
+      "feature_title_2": "Step 2: Add Your Materials",
+      "feature_description_2": "Upload your existing documents, presentations, and teaching resources to build a personalized content library for each class.",
+      "feature_title_3": "Step 3: Generate with AI",
+      "feature_description_3": "Use our AI tools to transform your materials into lesson plans, assessments, and activities tailored to your teaching style and curriculum.",
+
+      "demo_title": "See Maestre in Action",
+      "demo_description": "Watch how our platform transforms lesson planning, assessment creation, and classroom management in minutes",
+      "demo_video_unsupported": "Your browser does not support the video tag",
+
+      "tools_title": "Explore our tools",
+      "tools_description": "Discover our suite of intelligent educational tools designed to save time and enhance your learning outcomes",
+
+      "cta_title": "Ready to transform your teaching?",
+      "cta_description": "Join thousands of educators worldwide who are saving time and improving learning outcomes with Maestre's innovative tools",
+      "signup_button": "Sign up as a Teacher",
+      "learn_more_button": "Learn more",
+      "existing_acc_button": "Already have an account?",
+      "signin_button": "Sign in here",
+
+      "team_title": "Meet our team",
+      "team_description": "The brilliant minds behind Maestre's innovative educational platform",
+      "team_member_bio_1": "Spearheading the frontend development and user experience design for Maestre's educational tools",
+      "team_member_bio_2": "Leading the technical architecture and backend development of Maestre's innovative education platform",
+      "team_member_role_1": "Software Engineer, Project Manager",
+      "team_member_role_2": "Software Engineer, Analyst",
+
+      "testimonial_text": "Maestre has completely transformed my teaching workflow. I can create engaging lessons in minutes that used to take hours, and my students are more engaged than ever before",
+      "testimonial_author": "Rafael Pulido",
+      "testimonial_job": "High School Geography and History teacher, Seville",
+
+      "stay_up_to_date": "Stay up to date",
+      "stay_up_to_date_description": "Get the latest news, resources and updates about Maestre's educational tools",
+      "email_placeholder": "Enter your email",
+      "subscribe_button": "Subscribe",
+      
+      "footer_about": "About us",
+      "footer_legal": "Legal",
+      "footer_terms": "Terms of Service",
+      "footer_privacy": "Privacy Policy",
+      "footer_cookies": "Cookie Policy",      
+      "footer_products": "Products",
+      "footer_exam_creator": "Exam Creator",
+      "footer_test_creator": "Test Creator",
+      "footer_translator": "Translator",
+      "footer_planner": "Lesson Planner",
+      "footer_copyright": "© 2025 Maestre. Under MIT License.",
+      "footer_language": "Language:",
+      
+      "tool_category_assessment": "Assessment",
+      "tool_category_organization": "Organization",
+      "tool_category_language": "Language",
+      "tool_title_exams": "Create Engaging Exams",
+      "tool_title_tests": "Generate Tests Efficiently",
+      "tool_title_planner": "Plan Your Classes",
+      "tool_title_translator": "Translate Materials",
+      "tool_content_exams": "Create comprehensive exams with our intuitive exam maker. Features include:",
+      "tool_content_tests": "Streamline your test creation process with AI-powered assistance:",
+      "tool_content_planner": "Stay organized with our comprehensive planning tools:",
+      "tool_content_translator": "Break language barriers in your classroom:",
+      "tool_feature_exams_1": "Multiple question types",
+      "tool_feature_exams_2": "Automatic grading",
+      "tool_feature_exams_3": "Custom scoring rules",
+      "tool_feature_exams_4": "Question bank management",
+      "tool_feature_tests_1": "AI-generated questions",
+      "tool_feature_tests_2": "Smart difficulty adjustment",
+      "tool_feature_tests_3": "Topic-based generation",
+      "tool_feature_tests_4": "Instant test assembly",
+      "tool_feature_planner_1": "Visual class schedules",
+      "tool_feature_planner_2": "Curriculum mapping",
+      "tool_feature_planner_3": "Resource allocation",
+      "tool_feature_planner_4": "Progress tracking",
+      "tool_feature_translator_1": "Real-time translation",
+      "tool_feature_translator_2": "Multi-language support",
+      "tool_feature_translator_3": "Educational context awareness",
+      "tool_feature_translator_4": "Document translation",
+      
+      "try_now_button": "Try Now",
+      "auth_modal_title": "Authentication Required",
+      "auth_modal_description": "You need to be logged in to access this tool. Create an account to unlock all features.",
+      "auth_modal_cancel": "Cancel",
+      "auth_modal_signup": "Create Account",
+      "auth_modal_signin": "Sign In",
+      "profile_modal_title": "Complete Your Profile",
+      "profile_modal_description": "We need a few more details before you can access this tool. Your profile information helps us personalize your experience.",
+      "profile_modal_cancel": "Cancel",
+      "profile_modal_complete": "Complete Profile"
+    }
+}
