+++ frontend/messages/es.json
@@ -1,0 +1,98 @@
+{
+    "HomePage": {
+      "title": "MAESTRE",
+      "subtitle": "Reimaginando la educación con inteligencia",
+      "main_button": "Descubre nuestras herramientas de IA",
+      "main_description": "Maestre combina la inteligencia artificial con la pedagógía para crear herramientas que mejoran la forma de trabajar y aprender de los docentes",
+      "loading": "Cargando contenido...",
+
+      "feature_title_1": "Paso 1: Crear y Gestionar Clases",
+      "feature_description_1": "Configura tu entorno de aula virtual creando y organizando clases para diferentes asignaturas y grupos de estudiantes.",
+      "feature_title_2": "Paso 2: Añade tus Materiales",
+      "feature_description_2": "Sube tus documentos existentes, presentaciones y recursos didácticos para construir una biblioteca de contenido personalizada para cada clase.",
+      "feature_title_3": "Paso 3: Genera con IA",
+      "feature_description_3": "Utiliza nuestras herramientas de IA para transformar tus materiales en planes de lección, evaluaciones y actividades adaptadas a tu estilo de enseñanza y plan de estudios.",
+
+      "demo_title": "Descubre Maestre en acción",
+      "demo_description": "Mira cómo nuestra plataforma transforma la planificación de clases, la creación de examenes y la gestión del aula en minutos",
+      "demo_video_unsupported": "Tu navegador no soporta este formato de video",
+      
+      "tools_title": "Explora todas nuestras herramientas",
+      "tools_description": "Descubre nuestro catálogo de herramientas educativas inteligentes diseñadas para ahorrar tiempo y mejorar tus resultados de aprendizaje",
+      
+      "cta_title": "¿List@ para transformar tus clases?",
+      "cta_description": "Únete a miles de educadores que están ahorrando tiempo y mejorando los resultados de aprendizaje con las herramientas de IA de MAESTRE",
+      "signup_button": "Regístrate como Docente",
+      "learn_more_button": "Saber más",
+      "existing_acc_button": "¿Ya tienes una cuenta?",
+      "signin_button": "Inicia sesión aquí",
+      
+      "team_title": "Conoce a nuestro equipo",
+      "team_description": "Las mentes detrás de la innovadora plataforma educativa de Maestre",
+      "team_member_bio_1": "Liderando el desarrollo frontend y el diseño de experiencia de usuario para las herramientas educativas de Maestre",
+      "team_member_bio_2": "Dirigiendo la arquitectura técnica y el desarrollo backend de la innovadora plataforma educativa de Maestre",
+      "team_member_role_1": "Ingeniero de Software, Gerente de Proyecto",
+      "team_member_role_2": "Ingeniero de Software, Analista",
+
+      "testimonial_text": "Maestre ha transformado completamente mi forma de dar clases. Puedo planear clases atractivas en minutos cuando antes me llevaba horas",
+      "testimonial_author": "Rafael Pulido",
+      "testimonial_job": "Profesor de Geografía e Historia de Secundaria, Sevilla",
+
+      "stay_up_to_date": "Mantente actualizado",
+      "stay_up_to_date_description": "Recibe las últimas noticias, recursos y actualizaciones sobre las herramientas educativas de Maestre",
+      "email_placeholder": "Introduce tu email",
+      "subscribe_button": "Subscríbete",
+      
+      "footer_about": "Sobre Nosotros",
+      "footer_legal": "Legal",
+      "footer_terms": "Términos de Servicio", 
+      "footer_privacy": "Política de Privacidad",
+      "footer_cookies": "Política de Cookies",
+      "footer_products": "Productos",
+      "footer_exam_creator": "Generador de Exámenes",
+      "footer_test_creator": "Generador de Tests",
+      "footer_translator": "Traductor",
+      "footer_planner": "Planificador de Clases",
+      "footer_copyright": "© 2025 Maestre. Bajo MIT License.",
+      "footer_language": "Idioma:",
+      
+      "tool_category_assessment": "Evaluación",
+      "tool_category_organization": "Organización",
+      "tool_category_language": "Idioma",
+      "tool_title_exams": "Crea Exámenes Atractivos",
+      "tool_title_tests": "Genera Tests Eficientemente",
+      "tool_title_planner": "Planifica tus Clases",
+      "tool_title_translator": "Traduce Materiales",
+      "tool_content_exams": "Crea exámenes completos con nuestro intuitivo creador de exámenes. Características incluidas:",
+      "tool_content_tests": "Agiliza tu proceso de creación de tests con asistencia impulsada por IA:",
+      "tool_content_planner": "Mantente organizado con nuestras completas herramientas de planificación:",
+      "tool_content_translator": "Rompe las barreras del idioma en tu aula:",
+      "tool_feature_exams_1": "Múltiples tipos de preguntas",
+      "tool_feature_exams_2": "Calificación automática",
+      "tool_feature_exams_3": "Reglas de puntuación personalizadas",
+      "tool_feature_exams_4": "Gestión de banco de preguntas",
+      "tool_feature_tests_1": "Preguntas generadas por IA",
+      "tool_feature_tests_2": "Ajuste inteligente de dificultad",
+      "tool_feature_tests_3": "Generación basada en temas",
+      "tool_feature_tests_4": "Ensamblaje instantáneo de tests",
+      "tool_feature_planner_1": "Horarios visuales de clase",
+      "tool_feature_planner_2": "Mapeo curricular",
+      "tool_feature_planner_3": "Asignación de recursos",
+      "tool_feature_planner_4": "Seguimiento de progreso",
+      "tool_feature_translator_1": "Traducción en tiempo real",
+      "tool_feature_translator_2": "Soporte multilingüe",
+      "tool_feature_translator_3": "Conciencia del contexto educativo",
+      "tool_feature_translator_4": "Traducción de documentos",
+
+      "try_now_button": "Probar Ahora",
+      "auth_modal_title": "Autenticación Requerida",
+      "auth_modal_description": "Necesitas iniciar sesión para acceder a esta herramienta. Crea una cuenta para desbloquear todas las funciones.",
+      "auth_modal_cancel": "Cancelar",
+      "auth_modal_signup": "Crear Cuenta",
+      "auth_modal_signin": "Iniciar Sesión",
+      "profile_modal_title": "Completa Tu Perfil",
+      "profile_modal_description": "Necesitamos algunos detalles más antes de que puedas acceder a esta herramienta. La información de tu perfil nos ayuda a personalizar tu experiencia.",
+      "profile_modal_cancel": "Cancelar",
+      "profile_modal_complete": "Completar Perfil"
+    }
+}
