+++ frontend/next.config.js
@@ -3,3 +3,12 @@ module.exports = {
     domains: ['ui.aceternity.com'],
   },
 };
+
+const createNextIntlPlugin = require('next-intl/plugin');
+ 
+const withNextIntl = createNextIntlPlugin();
+ 
+/** @type {import('next').NextConfig} */
+const nextConfig = {};
+ 
+module.exports = withNextIntl(nextConfig);
