+++ frontend/src/app/[locale]/classrooms/[id]/page.js
