+++ frontend/src/app/[locale]/classrooms/page.js
