+++ frontend/src/app/[locale]/favicon.ico
