+++ frontend/src/app/[locale]/materials/page.js
