+++ frontend/src/app/[locale]/not-found.js
