+++ frontend/src/app/[locale]/profile/signup/page.js
