+++ frontend/src/app/[locale]/schools/new/page.js
