+++ frontend/src/app/[locale]/terms/page.js
