+++ frontend/src/app/[locale]/tools/exam-maker/components/ExamResultModal.jsx
