+++ frontend/src/app/[locale]/tools/exam-maker/components/FileUploadSection.jsx
