+++ frontend/src/app/[locale]/tools/exam-maker/components/LabelInputContainer.jsx
