+++ frontend/src/app/[locale]/tools/exam-maker/components/MaterialsModal.jsx
