+++ frontend/src/app/[locale]/tools/exam-maker/components/ScoringStyleSelector.jsx
