+++ frontend/src/app/[locale]/tools/exam-maker/utils/pdfUtils.js
