+++ frontend/src/app/[locale]/tools/exam-maker/utils/promptUtils.js
