+++ frontend/src/app/[locale]/tools/page.js
