+++ frontend/src/i18n/routing.js
@@ -1,0 +1,9 @@
+import {defineRouting} from 'next-intl/routing';
+ 
+export const routing = defineRouting({
+  // A list of all locales that are supported
+  locales: ['en', 'es'],
+ 
+  // Used when no locale matches
+  defaultLocale: 'es'
+});
