+++ frontend/src/middlewares.js
@@ -1,0 +1,16 @@
+import createMiddleware from 'next-intl/middleware';
+import {routing} from './i18n/routing';
+ 
+export default createMiddleware({
+  // Add the localePrefix setting to clarify how locales should be handled
+  localePrefix: 'always',
+  // Use the routing configuration from your i18n setup
+  ...routing
+});
+ 
+export const config = {
+  // Match all pathnames except for
+  // - … if they start with `/api`, `/trpc`, `/_next` or `/_vercel`
+  // - … the ones containing a dot (e.g. `favicon.ico`)
+  matcher: ['/((?!api|trpc|_next|_vercel|.*\\..*).*)']
+};
